+++ .angular/cache/19.0.7/benchmark-test/vite/deps/@angular_router.js
@@ -1,13 +1,16 @@
 import {
+  Title
+} from "./chunk-PQDH54U2.js";
+import "./chunk-6OXBDMVW.js";
+import {
   DOCUMENT,
   HashLocationStrategy,
   LOCATION_INITIALIZED,
   Location,
   LocationStrategy,
   PathLocationStrategy,
-  Title,
   ViewportScroller
-} from "./chunk-PICPI3V2.js";
+} from "./chunk-J3L2LTDM.js";
 import {
   APP_BOOTSTRAP_LISTENER,
   APP_INITIALIZER,
