+++ src/app/types/product.interface.ts
@@ -1,0 +1,11 @@
+export interface Product {
+  id: number;
+  title: string;
+  description: string;
+  category: string;
+  stock: number;
+  thumbnail: string;
+  images: string[];
+  rating: number;
+  price: number;
+}
